[GOHW-2]: переход на grpc

See merge request sirejja/Homework!3

+++ checkout/internal/services/cart/delete_from_cart.go
@@ -4,7 +4,7 @@ import (
 	"context"
 )
 
-func (c *Cart) DeleteFromCart(ctx context.Context, user int64, sku uint32, count uint16) error {
+func (c *Cart) DeleteFromCart(ctx context.Context, user int64, sku uint32, count uint32) error {
 	//op := "Cart.DeleteFromCart"
 	return nil
 }
